+++ src/services/taptitan/card_function/burst/barbed_morningstar.rs
@@ -15,7 +15,7 @@ pub fn get_proc_chance(card: &Card, _boss: &Boss) -> f64 {
     }
 }
 
-pub fn on_proc(card: &Card, boss: &mut Boss, target_part: BossPartName, damage: f64) -> u64 {
+pub fn on_proc(card: &Card, boss: &mut Boss, target_part: BossPartName, damage: f64) -> f64 {
     let burst_mult = card.skill.value_a.unwrap_or(1.0);
     let armor_damage_boost = card.skill.bonus_c.unwrap_or(0.0);
     let body_damage_boost = card.skill.bonus_d.unwrap_or(0.0);
@@ -32,7 +32,7 @@ pub fn on_proc(card: &Card, boss: &mut Boss, target_part: BossPartName, damage: 
         PartState::Body => 1.0 + body_damage_boost * armor_part_count(boss, max_bonus_parts) as f64,
         PartState::Skeleton => 1.0,
     };
-    let result_damage = (damage * burst_mult * bonus_mult).max(0.0) as u64;
+    let result_damage = (damage * burst_mult * bonus_mult).max(0.0);
 
     boss.on_hit_with_source(target_part, result_damage, DamageSource::Card(card.card_id));
     result_damage
